Broke out parsing and rendering of cell text for pdf builder so background can be rendered first. Fixes issue #4

+++ pdf/src/main/groovy/com/craigburke/document/builder/render/ParagraphRenderer.groovy
@@ -46,7 +46,7 @@ class ParagraphRenderer {
         }
     }
 
-    static void renderLine(Document document, ParagraphLine line, int renderStartX, RenderState renderState) {
+    static void renderLine(Document document, ParagraphLine line, float renderStartX, RenderState renderState) {
         PdfDocument pdfDocument = document.item
 
         if (renderState == RenderState.PAGE && pdfDocument.remainingPageHeight < line.lineSpacing) {
